backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.3
pydantic-settings==2.0.3
python-dotenv==1.0.0
aiohttp==3.9.0
apscheduler==3.10.4
pandas==2.1.3
numpy==1.24.3
httpx==0.25.1
sentence-transformers==2.2.2
torch==2.1.0
gymnasium==0.29.0
matplotlib==3.8.2